backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
gitpython==3.1.40
pydantic==2.5.0
openai==1.51.0
python-dotenv==1.0.0
requests==2.31.0